Fixed the Gamestop price and also changed the ProcessCheckoutInfo struct

+++ juiced.sitescripts/amazon/task.go
@@ -598,7 +598,7 @@ func (task *Task) PlaceOrder(startTime time.Time) (bool, enums.OrderStatus) {
 		ItemName:     task.TaskInfo.ItemName,
 		Sku:          task.TaskInfo.ASIN,
 		Retailer:     enums.Amazon,
-		Price:        task.CheckoutInfo.Price,
+		Price:        float64(task.CheckoutInfo.Price),
 		Quantity:     1,
 		MsToCheckout: time.Since(startTime).Milliseconds(),
 	})
